move sections to admindashboard and pass down
